Add Google Calendar integration
Refactor parser to output a generic Array of events instead of an
iCalendar text. This allows a simpler import of the events to Google
Calendar, which requires a custom syntax and does not accept iCalendar
through its API. The generic events could also be useful for adding the
events to a different calendar provider, such as Microsoft Outlook.

Rearrange repo structure to better separate Javascript code from HTML
and CSS.

Add the possibility to add the events to Google Calendar on the stand-
alone webpage. This allows the user to import the events faster. The app
can create a new calendar if necessary. Events are imported in parallel
with an insert operation each. Batch import seemed adequate, but Google
announced it will cease to work on August 12, 2020 and their Javascript
API hasn't been updated to solve the problem.