Fix output on clients without Europe/Rome timezone
The code always assumed the client used the Europe/Rome timezone. The assumption was too weak, as it was enough to have the automatic daylight saving disabled to get issues.
This commit refactors the structure of the code to allow the use of the Luxon time library. The website code is a single module (plus the Google API script), which is bundled by esbuild. A different iCalendar library has been used as well to easily add the required timezone info.